docs: reorganize scripts
