+++ scripts/build-manifest/fix-inherited-field-types.js
