+++ scripts/build-manifest/index.js
