+++ scripts/build-manifest/option-property-decorator.js
