+++ scripts/build-manifest/remove-internal-classes.js
