+++ scripts/build-manifest/remove-members.js
