+++ scripts/build-manifest/sort.js
